+++ src/app.ts
@@ -1,5 +1,5 @@
 import * as express from "express";
-import { Cat, CatType } from './app.model';
+import catsRouter from "./cats/cats.route";
 
 const app: express.Express = express();
 
@@ -13,67 +13,7 @@ app.use((req, res, next) => {
 // json middleware
 app.use(express.json());
 
-app.get('/cats/som', (req, res, next) => {
-    console.log("this is som middleware.");
-    next();
-});
-
-// READ 고양이 전체 데이터
-app.get('/cats', (req, res) => {
-    try {
-        const cats = Cat;
-        res.status(200).send({
-            success: true,
-            data: {
-                cats
-            }
-        })
-    } catch (error) {
-        res.status(400).send({
-            success: false,
-            error: error.message
-        })
-    }
-})
-
-// READ 특정 고양이 데이터 조회
-app.get('/cats/:id', (req, res) => {
-    try {
-        const params = req.params;
-        const cats = Cat.find((cat) => {
-            return cat.id === params.id;
-        })
-        
-        res.status(200).send({
-            success: true,
-            data: {
-                cats
-            }
-        })
-    } catch (error) {
-        res.status(400).send({
-            success: false,
-            error: error.message
-        })
-    }
-})
-
-// CREATE 새로운 고양이 추가
-app.post('/cats', (req, res) => {
-    try{
-        const data = req.body;
-        Cat.push(data);
-        res.status(200).send({
-            success: true,
-            data: { data }
-        })
-    } catch (error) {
-        res.status(400).send({
-            success: false,
-            error: error.message
-        })
-    }
-})
+app.use(catsRouter);
 
 // 404 middleware
 app.use((req, res, next) => {
